Beginning of the year commit, continue working on UI

+++ src/cyclic_iterator.cpp
@@ -10,6 +10,6 @@ CyclicIterator& CyclicIterator::operator--() {
     return *this;
 }
 
-int CyclicIterator::operator*() const {
+uint8_t CyclicIterator::operator*() const {
     return value;
 }
